# src/main/java/com/revolut/database/entity/Account.java
package com.revolut.database.entity;

import javax.persistence.*;
import java.math.BigDecimal;

/**
 * Represents bank account in system
 * since the system is quite simple also directly serializes to JSON in Controller layer
 */
@Entity
@Table(name = "account")
public class Account {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;
    private String name;
    private BigDecimal balance;

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public BigDecimal getBalance() {
        return balance;
    }

    public void setBalance(BigDecimal balance) {
        this.balance = balance;
    }
}