# src/main/java/com/revolut/database/entity/TransactionStatus.java
package com.revolut.database.entity;

/**
 * Represents status of transaction in our system
 * it is the only field containing processing information
 * if transaction fails system do not persist fail reason
 */
public enum TransactionStatus {
    SUCCESSFUL, UNSUCCESSFUL
}